DONE-2/METRICS-REDIS/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg
pymongo
redis
